+++ src/openServices/common/resolverOpenTools.js
@@ -7,6 +7,8 @@ import { phrasesResolvers } from '../phrases/phrasesResolvers';
 import { coexpressionResolver } from '../coexpressionService/coexpressionResolver';
 import { overviewsResolver } from '../overviewsService/overviews_resolver'
 import { regulonResolvers } from '../regulonService/regulon_resolver'
+import { sigmulonResolvers } from '../sigmulonService/sigmulon_resolver'
+import { htResolvers } from '../htService/ht_search_resolver'
 
 /** merges all resolver file and exports them to index */
 export const resolvers = mergeResolvers([
@@ -15,5 +17,6 @@ export const resolvers = mergeResolvers([
     operonResolvers, 
     coexpressionResolver,
     overviewsResolver,
-    regulonResolvers
+    regulonResolvers,
+    sigmulonResolvers
 ]);