+++ src/service_template/template_controller.js
@@ -1,0 +1,9 @@
+import {Template} from "./template_model"
+
+class templateController {
+    static async queryName() {
+        
+    }
+}
+
+export {templateController}